requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.4
fakeredis==2.21.1
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
influxdb-client==1.40.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==1.1.0
pytest==8.1.0
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
reactivex==4.0.4
redis==5.1.0b4
setuptools==69.1.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
